+++ example/node/countries-react/src/App.tsx
@@ -1,0 +1,6 @@
+import { CountriesPage } from "./features/countries/CountriesPage";
+
+export function App() {
+  return <CountriesPage />;
+}
+